All JS files included in index.php without errors 4/20
