+++ views/partials/common_js.php
@@ -1,8 +1,3 @@
-<!DOCTYPE html>
-<html>
-<body>
-<script src="https://ajax.googleapis.com/ajax/libs/jquery/2.1.3/jquery.min.js"></script>
-<script src="https://maxcdn.bootstrapcdn.com/bootstrap/3.3.7/js/bootstrap.min.js" integrity="sha384-Tc5IQib027qvyjSMfHjOMaLkfuWVxZxUPnCJA7l2mCWNIpG9mGCD8wGNIcPD7Txa" crossorigin="anonymous"></script>
 <script>
     $(document).ready(function() {
         $('.input-required').on('invalid', function() {
@@ -14,5 +9,3 @@
         });
     });
 </script>
-</body>
-</html>
